SSH_Homework/requirements.txt
djangorestframework==3.7.1
Markdown==2.6.9
django_filter==1.1.0
django==1.11.6